add crop visualisation to view

+++ Options/CropOptions.cs
@@ -11,29 +11,29 @@ namespace SampleSegmenter.Options
             set { SetProperty(ref _isEnabled, value); }
         }
 
-        private int _x;
-        public int X
+        private double _x;
+        public double X
         {
             get { return _x; }
             set { SetProperty(ref _x, value); }
         }
 
-        private int _y;
-        public int Y
+        private double _y;
+        public double Y
         {
             get { return _y; }
             set { SetProperty(ref _y, value); }
         }
 
-        private int _height;
-        public int Height
+        private double _height;
+        public double Height
         {
             get { return _height; }
             set { SetProperty(ref _height, value); }
         }
 
-        private int _width;
-        public int Width
+        private double _width;
+        public double Width
         {
             get { return _width; }
             set { SetProperty(ref _width, value); }
